Move mock classes to independent package
