+++ common/src/test/java/com/github/khanshoaib3/minecraft_access/test_utils/MockInterval.java
@@ -1,6 +1,8 @@
-package com.github.khanshoaib3.minecraft_access.utils.condition;
+package com.github.khanshoaib3.minecraft_access.test_utils;
 
-class MockInterval extends Interval {
+import com.github.khanshoaib3.minecraft_access.utils.condition.Interval;
+
+public class MockInterval extends Interval {
     boolean ready = false;
 
     public MockInterval(long lastRunTime, long delayInNanoTime) {
