+++ common/src/test/java/com/github/khanshoaib3/minecraft_access/test_utils/MockKeystrokeActionTest.java
@@ -1,5 +1,6 @@
-package com.github.khanshoaib3.minecraft_access.utils.condition;
+package com.github.khanshoaib3.minecraft_access.test_utils;
 
+import com.github.khanshoaib3.minecraft_access.utils.condition.Keystroke;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
